requirements.txt
fastapi==0.99.1
sqlalchemy==2.0.17
jira==3.5.2
uvicorn==0.22.0
sqlalchemy-cockroachdb==2.0.1
psycopg2==2.9.6
python-dateutil==2.8.2